WIP Send player messages from the UI thread
Fix for the following warning when AudioPlayer functions like
setAudioSource are called:

[ERROR:flutter/shell/common/shell.cc(1015)]
The 'com.ryanheise.just_audio.events....' channel sent a message from
native to Flutter on a non-platform thread. Platform channel messages
must be sent on the platform thread. Failure to do so may result in
data loss or crashes, and must be fixed in the plugin or application
code creating that channel...

WIP:

- At some point Flutter will support sending messages from any thread.
  Tracking issue: https://github.com/flutter/flutter/issues/93945
  So this is mostly a stopgap solution.

- Alternatively to the enable_shared_from_this pattern, posted
  messages could capture weak pointers to the event sink. This would
  behave subtly differently when the sink pointer changes between
  posting the message and executing the handler (i.e., upon any
  change, the handler would see an empty pointer. This is possibly
  more correct).